Conflict resolution improvements
It turns out a node may actually receive *its own* packet in case
of a conflict, and we don't appear to be able to catch the collision
in time to prevent our own transmission. so just flip a coin to
decide whether to move slots when we hear anything using our slot.

To facilitate this, add a small random delay to packet transmission
times so a collision is more likely to be detected, and increase
the resolution of the delay field in the packet.